Step 07: separate classes
