+++ firebase-get-to-know-flutter/step_07/test/widget_test.dart
@@ -1,4 +1,5 @@
 import 'package:flutter_test/flutter_test.dart';
+import 'package:gtk_flutter/app_state.dart';
 import 'package:gtk_flutter/main.dart';
 import 'package:provider/provider.dart';
